+++ utils/importTypes/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+__author__ = 'XaviTorello'
+
+from f1 import process_F1file
+from q1 import process_Q1file
+
